Be more defensive
When removing snapshot, stopping a vm, or removing a vm, engine may fail
with crazy errors like:

    ovirtsdk4.Error: The response content type 'text/html;
    charset=iso-8859-1' isn't the expected XML

This is likely a bug in engine http server.

In other cases engine fail because the operation is not valid:

    ovirtsdk4.Error: Fault reason is "Operation Failed". Fault detail is
    "[Cannot remove Snapshot. The operation is only allowed for regular
    snapshots.]". HTTP response code is 409.

This looks like a race in engine, after a snapshot reports status OK.

Another case is validation error - the operation is not valid at this
time:

    Fault reason is "Operation Failed". Fault detail is "[Cannot remove
    VM. The VM is performing an operation on a Snapshot. Please wait for
    the operation to finish, and try again.]". HTTP response code is
    409.

These errors are communicated via generic sdk.Error, so we don't have a
good way to handle them. Aborting the operation or polling is not
useful since the entire runner may fail if we fail to remove a vm at the
end of the test.

Try to be more robust by retrying operations 10 times after sdk.Error.
If the issue is a race or temporary failure, the next retry will
succeed. If the issue is validation, the state of the system may change
and the operation may succeed.

During polling, log a warning about the sdk.Error and continue polling.
Hopefully the next polling will succeed. In the worst case we will fail
until we time out.