Fix waiting for vm and snapshot
Waiting for entity when during creation or removal does not work or not
efficient. The only way to do this is using jobs and correlation id.

Now we wait for the jobs related to create vm, create snapshot, or
remove snapshot operations. We detect failures by checking the job
status.

In the case when creating a vm, creating a snapshot or removing a
snapshot fail, we can abort the current iteration immediately instead of
waiting few minutes for timeout.

Fixes #1