Don't wait forever for removing vm
We use a unique vm name for every iteration, so if removing vm fail we
can leave the vm as is and continue. The user will have to clean up the
vm after the tests.

Without this remove vm can get stuck for hours waiting for vm removal
that will never happen.