Simplify cleanup flow
The actual issue when stopping vm is that engine fails with sdk.Error if
the vm is not running. Checking the vm status before stopping is racy
but avoids this issue.

When removing a vm, engine may fail if the vm cannot be removed. In this
case we can leave the vm as is and start the next iteration.