Increase timeouts
I'm testing now 8 vms with 16 data disks, committing 32 GiB for every
delete snapshot. With this load, everything becomes much slower so we
need longer timeouts. Since we moved to waiting for jobs, we detect
errors early, and using long timeouts does not slow down the test unless
the system is under extreme load.

The comments about failing the test if vm cannot be removed are not
relevant now, remove them.