Check data center health
When SPM fail in a bad way, engine try to move the SPM to another host.
During this time the data center is UNKNOWN and every storage operation
will fail. Current code fails and retry immediately, possibly failing
all iterations.

Here is an example run when 55 iterations failed in 27 seconds before
the SPM was recovered:

2021-03-20 12:01:57,338 INFO    (run/6) Creating snapshot for vm delete-snspshot-vm-6-45
2021-03-20 12:01:57,757 ERROR   (run/6) Iteration 45 errored
2021-03-20 12:01:57,773 INFO    (run/6) Stopping vm delete-snspshot-vm-6-45
2021-03-20 12:02:19,937 DEBUG   (run/6) VM delete-snspshot-vm-6-45 is down
2021-03-20 12:02:19,937 INFO    (run/6) VM delete-snspshot-vm-6-45 stopped in 22 seconds
2021-03-20 12:02:19,937 INFO    (run/6) Removing vm delete-snspshot-vm-6-45
2021-03-20 12:02:20,379 ERROR   (run/6) Error tearing down
2021-03-20 12:02:20,380 INFO    (run/6) Iteration 45 completed in 463 seconds
2021-03-20 12:02:20,380 INFO    (run/6) Iteration 46 started
2021-03-20 12:02:20,380 INFO    (run/6) Creating vm delete-snspshot-vm-6-46
2021-03-20 12:02:20,861 ERROR   (run/6) Iteration 46 errored
2021-03-20 12:02:20,861 INFO    (run/6) Iteration 46 completed in 0 seconds
...
2021-03-20 12:02:44,945 INFO    (run/6) Iteration 99 started
2021-03-20 12:02:44,945 INFO    (run/6) Creating vm delete-snspshot-vm-6-99
2021-03-20 12:02:45,409 ERROR   (run/6) Iteration 99 errored
2021-03-20 12:02:45,409 INFO    (run/6) Iteration 99 completed in 0 seconds
2021-03-20 12:02:47,452 INFO    (run/6) Finished

Only one runner was affected, but the next 8 hours of this test run with
7 runners instead of 8.

Fix by checking data center status before setup() and teardown(), and
waiting until the data center is UP before continuing.