Add backup stress test
Create a configuration and ansible setup for backup vm template.  We
start with a minimal vm using single OS disk. More disks will be added
later, or maybe we will use multiple templates.

The new test is mostly a copy of the delete snapshot test. When the
test is finished, we can start thinking how to reuse test code better.

The imagetransfer module was borrowed form the python sdk examples
helpers with minimal changes. The backup is based on the backup_vm.py
example.