backup: Modify guest data between backups
Previously the incremental backup was almost empty since the only change
from the full backup was creation of one file. Real backups contain data
modified during one day.

Now we write 1g of zeroes to the image during the full backup. This data
will be merged when stopping the backup, and will be included in the
incremental backup.

During the incremental backup we delete the data file created during the
full backup, and use fstrim the convert the deleted data to a hole. Then
we write more data that will be merged when we stop the backup.

This adds more I/O load to the test, both when downloading the backups
and when deleting snapshots. This load can slow down the hosts and
reveal timing issues.