backup: Add start stop backup timeout
Testing shows that starting a backup can get stuck waiting for
completion of the snapshot. We also know that deleting a snapshot when
stopping a backup may get stuck.

In the last test run, the second iteration was stuck for 9 hours until I
destroyed the VM manually:

    2022-03-11 14:22:48,398 INFO    (run/4) Iteration 1 completed in 32633 seconds

And then the next iteration started:

    2022-03-11 14:22:48,398 INFO    (run/4) Iteration 2 started

This is not useful. We want test run to terminate in predictable time,
and mark this run as a failure, which is the way real backup application
will treat such event.

Add timeouts to start_backup() and stop_backup(). On timeout we will
fail the iteration, cleanup, and start a new iteration.