backup: Support interrupting the test
Previously we lost the stats if a test was interrupted. Now stats are
log in this case. This is useful when you discover that a test
configuration is much slower than expected and want to abort the test
early.

When a test is interrupted, we don't do any cleanup - the system is left
as is and should be cleaned manually. This allow inspecting a running
system.

Example aborted run:

    2022-03-11 16:32:52,816 INFO    (MainThread) Test interrupted,
    aborting test without cleanup

    2022-03-11 16:32:52,816 INFO    (MainThread) 16 full backups, 35
    incremental backups, 0 failed, 35 passed, 1 errored in 14.7 minutes