backup: Log download time
For fast storage and small downloads, the time to download the data is
much smaller than the time to finalize the transfer. We included the
time to create and finalize a transfer in the download time, which makes
it harder to understand how time is spent.

Since imagetransfer helpers already already log and time creating a
transfer and finalizing it, we just need to log the start and finish of
the download part.