ci: add pylint job
Add pylint.yml to have basic linting CI job.
The workflow is the official GH pylint template
with some extra installs to satisfy dependencies.

This requires installing dependiencies on each run,
but is a good starting point for the new workflow.

Use an autogenerated pylintrc configuration
that passes with current code.

Signed-off-by: Albert Esteve <aesteve@redhat.com>